New mode: Re-docking
Allows an engineer to select a docked vessel and move it to a different
docking port. The same can be done via detach/move action but with
re-dock action there is no need to snipe the right docking port - the
action will find it automatically. Also, in the re-docking mode
"attachment" it's a default  and the only possible mode.