Change a fallback approach when no attach nodes found
When the pointer is being created the fallback options on "no attach
node" event are very limited. Throwing an exception may be a good idea
for the development but it's not as good for the end user. So, if it
happen just pick whatever node we can, and hope the caller code has done
enough to prevent this situation to happen. Not the case as for now,
though.