Fix #118
Instead of guessing which pickup module triggered the event just request
the module form the active vessel. Since the action is served from a
mouse event handler in 99% of cases it's an EVA kerbal.

Leave 1% for other modders who may decide to simulate mouse actions (BAD
idea!), and just don't play any sound in this case.