Add bottom attachment node to container mount
This adds a small attachment node to the bottom of the container mount,
to allow for precise positioning when attaching it in the field.  In
particular, it fits nicely on top of the small structural panel, on the
side of the MKS MultiTruss or MiniTruss, and on the bed of the USI Akita
rover -- all of which have attachment nodes.

(Of course, the container can already be attached directly to those
nodes, but that requires an engineer with a screwdriver.  This change
effectively allows an engineer to make any attachment node accessible to
non-engineers for container placement.)