Fix #320
This is actually an expected behavior. The KIS items that are marked as
`allowPartAttach = Disabled` cannot be attached ort detached in EVA.
However, for teh consistency, such parts hsould also be restricted to be
attached in the editor. Otherwise, it's possible to create a vessel that
cannot be disassembled in EVA via KIS.