Raycast attach node to check for colliders
The ray cast is used to check whether then attach node is inside a
collider on the part to be dropped, and if so find the offset needed to
avoid dropping the part with the collider inside another collider, thus
preventing the part from flying off in low gravity. Of course, the
offset will not be sufficient of the collider surface is not
perpendicular to the attach node, or if the overall part collider setup
is concave around the attach node.

The computed offset works as a default value for the user-adjustable
offset.