Cope with the inconsistency of surface attach nodes
It seems the in/out direction of the surface attach node makes no
difference to part orientation during editor or field placement (must be
some sort of cancellation), thus there's no consistency between parts
(even stock) because there was nothing to enforce it. Therefore, it's
necessary to test both directions. The smaller of the two offsets is
used on the assumption that the node is at least positioned correctly.